Task 6916 - CreateNotification manager to schedule appropriate notifcations - Part 2
Created AlarmNotificationManager - this is a singleton that can be called into from anywhere in the Mimicker code.  It is called after any alarm changes are made  through 1) scheduling by UI or after reboot 2) After alarms have completed ringing and 3) when changes are made to the notification settings.
Added two new settings - 1) Enable Notifications - this enables display of the next alarm that will ring in the notification area. This is enabled by default. 2) Improve reliability - This setting is dependant on setting 1) and it acquires the partial wakelock so that the CPU will not go to sleep.  This is  disabled by default as this setting will have an impact on battery life.